fix: duplicate config
commit-id:6c1b742d
