+++ crates/gateway/src/config_test.rs
@@ -1,6 +1,7 @@
-use crate::GatewayConfig;
-use assert_matches::assert_matches;
+use crate::config::GatewayConfig;
+
 use clap::Command;
+
 use papyrus_config::loading::load_and_process_config;
 use std::fs::File;
 use std::path::Path;
@@ -18,23 +19,12 @@ fn get_config_file(file_name: &str) -> Result<GatewayConfig, papyrus_config::Con
 fn test_valid_config() {
     // Read the valid config file and validate its content.
     let expected_config = GatewayConfig {
-        bind_address: String::from("0.0.0.0:8080"),
+        ip: "0.0.0.0".parse().unwrap(),
+        port: 8080,
     };
+
     let loaded_config = get_config_file(CONFIG_FILE).unwrap();
 
     assert!(loaded_config.validate().is_ok());
     assert_eq!(loaded_config, expected_config);
-}
-
-#[test]
-fn test_address_validator() {
-    // Read the valid config file and check that the validator finds no errors.
-    let mut config = get_config_file(CONFIG_FILE).unwrap();
-    assert!(config.validate().is_ok());
-
-    // Invalidate the bind address and check that the validator finds an error.
-    config.bind_address = String::from("bad_address");
-    assert_matches!(config.validate(), Err(e) => {
-        assert_eq!(e.field_errors()["bind_address"][0].code, "Invalid Socket address.");
-    });
 }
